# backend/tests/contract/openapi.json
{
 "components": {
  "schemas": {
...
   "AssetClassOut": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name"
    ],
    "title": "AssetClassOut",
    "type": "object"
   },
...
   "AssetDetailsWithPosition": {
    "properties": {
     "acc_return": {
      "default": 0.0,
      "title": "Acc Return",
      "type": "number"
     },
     "asset_type": {
      "$ref": "#/components/schemas/AssetTypeOut"
     },
     "asset_type_id": {
      "title": "Asset Type Id",
      "type": "integer"
     },
     "average_price": {
      "default": 0.0,
      "title": "Average Price",
      "type": "number"
     },
     "cagr": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cagr"
     },
     "etf": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ETFOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "exchange_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Exchange Id"
     },
     "fii": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FIIOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "fixed_income": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FixedIncomeOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "fund": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/InvestmentFundOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "price": {
      "default": 0.0,
      "title": "Price",
      "type": "number"
     },
     "quantity": {
      "default": 0.0,
      "title": "Quantity",
      "type": "number"
     },
     "stock": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/StockOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "ticker": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ticker"
     },
     "treasury_bond": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/TreasuryBondOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "twelve_months_return": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Twelve Months Return"
     },
     "value": {
      "default": 0.0,
      "title": "Value",
      "type": "number"
     }
    },
    "required": [
     "id",
     "ticker",
     "name",
     "asset_type_id",
     "asset_type"
    ],
    "title": "AssetDetailsWithPosition",
    "type": "object"
   },
...
   "AssetTypeOut": {
    "properties": {
     "asset_class": {
      "$ref": "#/components/schemas/AssetClassOut"
     },
     "asset_class_id": {
      "title": "Asset Class Id",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "short_name": {
      "title": "Short Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "asset_class_id",
     "short_name",
     "name",
     "asset_class"
    ],
    "title": "AssetTypeOut",
    "type": "object"
   },
...
   "ETFOut": {
    "properties": {
     "asset_id": {
      "title": "Asset Id",
      "type": "integer"
     },
     "segment": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ETFSegmentOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "segment_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Segment Id"
     }
    },
    "required": [
     "asset_id",
     "segment_id",
     "segment"
    ],
    "title": "ETFOut",
    "type": "object"
   },
   "ETFSegmentOut": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name"
    ],
    "title": "ETFSegmentOut",
    "type": "object"
   },
...
   "FIIOut": {
    "properties": {
     "asset_id": {
      "title": "Asset Id",
      "type": "integer"
     },
     "segment": {
      "$ref": "#/components/schemas/FIISegmentOut"
     },
     "segment_id": {
      "title": "Segment Id",
      "type": "integer"
     }
    },
    "required": [
     "asset_id",
     "segment_id",
     "segment"
    ],
    "title": "FIIOut",
    "type": "object"
   },
...
   "FIISegmentOut": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "type": {
      "$ref": "#/components/schemas/FIITypeOut"
     },
     "type_id": {
      "title": "Type Id",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "name",
     "type_id",
     "type"
    ],
    "title": "FIISegmentOut",
    "type": "object"
   },
   "FIITypeOut": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name"
    ],
    "title": "FIITypeOut",
    "type": "object"
   },
...
   "FixedIncomeOut": {
    "properties": {
     "asset_id": {
      "title": "Asset Id",
      "type": "integer"
     },
     "fee": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fee"
     },
     "fixed_income_type": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FixedIncomeTypeOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "fixed_income_type_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fixed Income Type Id"
     },
     "index": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/IndexOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "index_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Index Id"
     },
     "maturity_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Maturity Date"
     }
    },
    "required": [
     "asset_id",
     "index_id",
     "fixed_income_type_id",
     "maturity_date",
     "fee",
     "fixed_income_type",
     "index"
    ],
    "title": "FixedIncomeOut",
    "type": "object"
   },
...
   "FixedIncomeTypeOut": {
    "properties": {
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name",
     "description"
    ],
    "title": "FixedIncomeTypeOut",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "IndexCurrencyOut": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "code",
     "name"
    ],
    "title": "IndexCurrencyOut",
    "type": "object"
   },
   "IndexOut": {
    "properties": {
     "currency": {
      "$ref": "#/components/schemas/IndexCurrencyOut"
     },
     "currency_id": {
      "title": "Currency Id",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "short_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Short Name"
     },
     "symbol": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Symbol"
     }
    },
    "required": [
     "id",
     "name",
     "short_name",
     "symbol",
     "currency_id",
     "currency"
    ],
    "title": "IndexOut",
    "type": "object"
   },
   "InvestmentFundOut": {
    "properties": {
     "anbima_category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Anbima Category"
     },
     "anbima_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Anbima Code"
     },
     "anbima_code_class": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Anbima Code Class"
     },
     "asset_id": {
      "title": "Asset Id",
      "type": "integer"
     },
     "legal_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Legal Id"
     }
    },
    "required": [
     "asset_id",
     "legal_id",
     "anbima_code",
     "anbima_code_class",
     "anbima_category"
    ],
    "title": "InvestmentFundOut",
    "type": "object"
   },
...
   "StockOut": {
    "properties": {
     "asset_id": {
      "title": "Asset Id",
      "type": "integer"
     },
     "country": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Country"
     },
     "industry": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Industry"
     },
     "sector": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sector"
     }
    },
    "required": [
     "asset_id",
     "sector",
     "country",
     "industry"
    ],
    "title": "StockOut",
    "type": "object"
   },
...
   "TreasuryBondOut": {
    "properties": {
     "asset_id": {
      "title": "Asset Id",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "maturity_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Maturity Date"
     },
     "type": {
      "$ref": "#/components/schemas/TreasuryBondTypeOut"
     },
     "type_id": {
      "title": "Type Id",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "asset_id",
     "type_id",
     "maturity_date",
     "type"
    ],
    "title": "TreasuryBondOut",
    "type": "object"
   },
   "TreasuryBondTypeOut": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "code",
     "name",
     "description"
    ],
    "title": "TreasuryBondTypeOut",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "auth/jwt/login"
     }
    },
    "type": "oauth2"
   }
  }
 },
 "info": {
  "description": "API for portfolio consolidation",
  "title": "My Stonks API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/portfolio/position/{portfolio_id}": {
   "get": {
    "operationId": "get_portfolio_position_portfolio_position__portfolio_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "most_recent",
      "required": false,
      "schema": {
       "default": true,
       "title": "Most Recent",
       "type": "boolean"
      }
     },
     {
      "in": "query",
      "name": "group_by_broker",
      "required": false,
      "schema": {
       "default": false,
       "title": "Group By Broker",
       "type": "boolean"
      }
     },
     {
      "in": "query",
      "name": "asset_id",
      "required": false,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Portfolio Position",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/analysis": {
   "get": {
    "operationId": "get_portfolio_analysis_portfolio_position__portfolio_id__analysis_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Portfolio Analysis",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/asset-type/{asset_type_id}/analysis": {
   "get": {
    "operationId": "get_asset_type_analysis_portfolio_position__portfolio_id__asset_type__asset_type_id__analysis_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "path",
      "name": "asset_type_id",
      "required": true,
      "schema": {
       "title": "Asset Type Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Asset Type Analysis",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/asset-type/{asset_type_id}/returns": {
   "get": {
    "operationId": "get_asset_type_returns_portfolio_position__portfolio_id__asset_type__asset_type_id__returns_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "path",
      "name": "asset_type_id",
      "required": true,
      "schema": {
       "title": "Asset Type Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Asset Type Returns",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/asset/{asset_id}/analysis": {
   "get": {
    "operationId": "get_asset_analysis_portfolio_position__portfolio_id__asset__asset_id__analysis_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "path",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Asset Analysis",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/asset/{asset_id}/details": {
   "get": {
    "operationId": "get_asset_details_portfolio_position__portfolio_id__asset__asset_id__details_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "path",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AssetDetailsWithPosition"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Asset Details",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/asset/{asset_id}/returns": {
   "get": {
    "operationId": "get_asset_returns_portfolio_position__portfolio_id__asset__asset_id__returns_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "path",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "start_date",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start Date"
      }
     },
     {
      "in": "query",
      "name": "end_date",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "End Date"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Asset Returns",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/category/returns": {
   "get": {
    "operationId": "get_category_returns_portfolio_position__portfolio_id__category_returns_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "category_id",
      "required": false,
      "schema": {
       "title": "Category Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "most_recent",
      "required": false,
      "schema": {
       "default": false,
       "title": "Most Recent",
       "type": "boolean"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Category Returns",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/category/{category_id}/analysis": {
   "get": {
    "operationId": "get_category_analysis_portfolio_position__portfolio_id__category__category_id__analysis_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "path",
      "name": "category_id",
      "required": true,
      "schema": {
       "title": "Category Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Category Analysis",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/patrimony_evolution": {
   "get": {
    "operationId": "get_patrimony_evolution_portfolio_position__portfolio_id__patrimony_evolution_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "asset_id",
      "required": false,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "asset_type_id",
      "required": false,
      "schema": {
       "title": "Asset Type Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "asset_type_ids",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "items": {
          "type": "integer"
         },
         "type": "array"
        },
        {
         "type": "null"
        }
       ],
       "title": "Asset Type Ids"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Patrimony Evolution",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
  "/portfolio/position/{portfolio_id}/returns": {
   "get": {
    "operationId": "get_portfolio_returns_portfolio_position__portfolio_id__returns_get",
    "parameters": [
     {
      "in": "path",
      "name": "portfolio_id",
      "required": true,
      "schema": {
       "title": "Portfolio Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Portfolio Returns",
    "tags": [
     "Portfolio Position"
    ]
   }
  },
...
 }
}